Create a random 1-10 printer
Prints the numbers 1-10 in pseudo random order.

Adds two tests. One tests whether all of 1-10 appears in the array. The
second tests whether the numbers are actually in random order. The
possibility of it being random but also being in order is there, however
it is low. Just for good measure I added in a nested check to lower the
probability of that ocurring.